+++ @Alyx/Alyx.m
@@ -24,7 +24,7 @@ classdef Alyx
   
   properties
     % URL to the Alyx database
-    BaseURL char = 'https://alyx.cortexlab.net'
+    BaseURL char = 'https://test.alyx.internationalbrainlab.org'
     % Set the local directory for saving queued Alyx commands, create if needed
     QueueDir char = 'C:\localAlyxQueue'
     % Set whether input dialogs should appear, e.g. login window
@@ -113,9 +113,11 @@ classdef Alyx
     % Checks for and uploads queued data to Alyx
     [data, statusCode] = flushQueue(obj)
     % Recovers the full filepath of a file on the repository, given the datasetURL
-    [fullPath, exists] = getFile(obj, eid, type)
+    [fullPath, exists] = getFile(obj, eid, type, remoteOnly)
     % Query the database for a list of sessions
     [sessions, eids] = getSessions(obj, varargin)
+    % Return an experiment reference string, given a session url/eid
+    ref = getExpRef(obj, eid)
     % Lists recorded subjects
     subjects = listSubjects(obj, stock, alive, sortByUser)
     % Returns the file path where you can find a specified file
@@ -136,7 +138,6 @@ classdef Alyx
     narrative = updateNarrative(obj, subject, comments, endpoint)
     % Return the instance of Alyx as a struct
     s = saveobj(obj)
-    % Validate Base URL string
   end
   
   methods (Access = private)
@@ -159,6 +160,8 @@ classdef Alyx
     updateSessions(obj, subjects, register)
     % Load an Alyx object from a struct
     obj = loadobj(s)
+    % Return eid part of url
+    eid = url2eid(url)
   end
   
 end