+++ @Alyx/expFilePath.m
@@ -38,7 +38,6 @@ function [fullpath, filename, fileID, records] = expFilePath(obj, varargin)
 %
 %   TODO:
 %     - Exists flag
-%     - Replace functionality of dat.expFilePath?
 %
 % Part of Alyx
 
@@ -46,6 +45,10 @@ function [fullpath, filename, fileID, records] = expFilePath(obj, varargin)
 
 % Validate input
 assert(nargin > 2, 'Error: Not enough arguments supplied.')
+
+% Flag for searching by session start time, rather than dataset created
+% time (see below)
+strictSearch = true;
 
 parsed = regexp(varargin{1}, dat.expRefRegExp, 'tokens');
 if isempty(parsed) % Subject, not ref
@@ -63,19 +66,21 @@ else % Ref, not subject
 end
 
 % Check date
-if ~ischar(expDate)||(ischar(expDate)&&length(expDate)==8)
+if ~ischar(expDate)
   expDate = datestr(expDate, 'yyyy-mm-dd');
-elseif ischar(expDate)&&length(expDate)==19
-  expDate = datestr(expDate, 'yyyy-mm-ddTHH:MM:SS');
+elseif ischar(expDate) && length(expDate) > 10
+  expDate = expDate(1:10);
 end
 
 if length(varargin) > 1 % Repository location defined
- location = varargin{2};
- % Validate repository
- repos = catStructs(obj.getData('data-repository')); %FIXME Datasets now returned as structs by default
- idx = strcmpi(location, {repos.name});
- assert(any(idx), 'Error: ''%s'' is an invalid data set type', location)
- location = repos(idx).name; % Ensures correct case
+  user = varargin{1};
+  location = varargin{2};
+  % Validate repository
+  repos = catStructs(obj.getData('data-repository'));
+  idx = strcmpi(location, {repos.name});
+  assert(any(idx), 'Alyx:expFilePath:InvalidType', ...
+    'Error: ''%s'' is an invalid data set type', location)
+  location = repos(idx).name; % Ensures correct case
 elseif ~isempty(varargin)
   user = varargin{1};
   location = [];
@@ -85,19 +90,48 @@ else
 end
 
 % Validate type
-dataSets = catStructs(obj.getData('dataset-types')); %FIXME Datasets now returned as structs by default
+dataSets = catStructs(obj.getData('dataset-types'));
 idx = strcmpi(type, {dataSets.name});
-assert(any(idx), 'Error: ''%s'' is an invalid data set type', type)
+assert(any(idx), 'Alyx:expFilePath:InvalidType', ...
+  'Error: ''%s'' is an invalid data set type', type)
 type = dataSets(idx).name; % Ensures correct case
 
 % Construct the endpoint
-endpoint = sprintf('/datasets?subject=%s&date=%s&experiment_number=%s&dataset_type=%s&created_by=%s',...
-  subject, expDate, num2str(seq), type, user);
+% FIXME: datasets endpoint filters no longer work
+% @body because of this we must make a seperate query to obtain the
+% datetime.  Querying the sessions takes around 3 seconds.  Otherwise we
+% filter by created time under the assumption that the dataset was created
+% on the same day as the session.  See https://github.com/cortex-lab/alyx/issues/601
+if strictSearch
+  endpoint = sprintf(['/datasets?'...
+    'subject=%s&'...
+    'experiment_number=%s&'...
+    'dataset_type=%s&'...
+    'created_by=%s'],...
+    subject, num2str(seq), type, user);
+  records = obj.getData(endpoint);
+  if ~isempty(records)
+    sessions = obj.getSessions(obj.url2eid({records.session}));
+    records = records(floor(obj.datenum({sessions.start_time})) == datenum(expDate));
+  end
+else
+  endpoint = sprintf(['/datasets?'...
+    'subject=%1$s&'...
+    'experiment_number=%2$s&'...
+    'dataset_type=%3$s&'...
+    'created_by=%4$s&'...
+    'created_datetime_gte=%5$s&'...
+    'created_datetime_lte=%5$s'],...
+    subject, num2str(seq), type, user, expDate);
+  records = obj.getData(endpoint);
+end
+% Construct the endpoint
+% endpoint = sprintf('/datasets?subject=%s&date=%s&experiment_number=%s&dataset_type=%s&created_by=%s',...
+%   subject, expDate, num2str(seq), type, user);
+% records = obj.getData(endpoint);
 
-% Get data
-records = obj.getData(endpoint);
 if ~isempty(records)
-  data = catStructs(records); %FIXME Datasets now returned as structs by default
+  data = catStructs(records);
   fileRecords = catStructs([data(:).file_records]);
 else
   fullpath = [];
@@ -108,12 +142,16 @@ end
 
 if ~isempty(location)
   % Remove records in unwanted repo locations
-  idx = contains({fileRecords.data_repository_path}, location);
-  data = data(idx);
+  idx = strcmp({fileRecords.data_repository}, location);
   fileRecords = fileRecords(idx);
 end
 
-fullpath = strcat({fileRecords.data_repository_path}, {fileRecords.relative_path});
+% Get the full paths
+mkPath = @(x) iff(isempty(x.data_url), ... % If data url not present
+  [x.data_repository_path x.relative_path], ... % make path from repo path and relative path
+  x.data_url); % otherwise use data_url field
+% Make paths
+fullpath = arrayfun(mkPath, fileRecords, 'uni', 0);
 filename = {data.name};
 fileID = {fileRecords.id};